+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/test/feature_counter/blocs/counter_bloc_test.dart
@@ -5,8 +5,9 @@
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
-import 'package:{{project_name}}/feature_counter/blocs/counter_bloc.dart';
+import 'package:{{project_name}}/base/models/count.dart';
 import 'package:{{project_name}}/base/repositories/counter_repository.dart';
+import 'package:{{project_name}}/feature_counter/blocs/counter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
@@ -25,7 +26,10 @@ void main() {
   group('CounterBloc tests', () {
     rxBlocTest<CounterBlocType, int>(
       'Initial state',
-      build: () async => CounterBloc(repo),
+      build: () async {
+        when(repo.getCurrent()).thenAnswer((_) async => Count(0));
+        return CounterBloc(repo);
+      },
       state: (bloc) => bloc.states.count,
       expect: [0],
     );
@@ -33,7 +37,8 @@ void main() {
     rxBlocTest<CounterBlocType, int>(
       'Incrementing value',
       build: () async {
-        when(repo.increment()).thenAnswer((_) async => 1);
+        when(repo.getCurrent()).thenAnswer((_) async => Count(0));
+        when(repo.increment()).thenAnswer((_) async => Count(1));
         return CounterBloc(repo);
       },
       act: (bloc) async => bloc.events.increment(),
@@ -44,7 +49,8 @@ void main() {
     rxBlocTest<CounterBlocType, int>(
       'Decrementing value',
       build: () async {
-        when(repo.decrement()).thenAnswer((_) async => -1);
+        when(repo.getCurrent()).thenAnswer((_) async => Count(0));
+        when(repo.decrement()).thenAnswer((_) async => Count(-1));
         return CounterBloc(repo);
       },
       act: (bloc) async => bloc.events.decrement(),
@@ -55,8 +61,9 @@ void main() {
     rxBlocTest<CounterBlocType, int>(
       'Increment and decrement value',
       build: () async {
-        when(repo.increment()).thenAnswer((_) async => 1);
-        when(repo.decrement()).thenAnswer((_) async => 0);
+        when(repo.getCurrent()).thenAnswer((_) async => Count(0));
+        when(repo.increment()).thenAnswer((_) async => Count(1));
+        when(repo.decrement()).thenAnswer((_) async => Count(0));
         return CounterBloc(repo);
       },
       act: (bloc) async {
@@ -70,8 +77,9 @@ void main() {
     rxBlocTest<CounterBlocType, String>(
       'Error handling',
       build: () async {
+        when(repo.getCurrent()).thenAnswer((_) async => Count(0));
         when(repo.increment()).thenAnswer(
-          (_) => Future.error('test error msg'),
+              (_) => Future.error('test error msg'),
         );
         return CounterBloc(repo);
       },
@@ -86,7 +94,8 @@ void main() {
     rxBlocTest<CounterBlocType, bool>(
       'Loading state',
       build: () async {
-        when(repo.increment()).thenAnswer((_) async => 1);
+        when(repo.getCurrent()).thenAnswer((_) async => Count(0));
+        when(repo.increment()).thenAnswer((_) async => Count(1));
         return CounterBloc(repo);
       },
       act: (bloc) async {